Initial commit of starting prototype for code generation library
The Makefile for this is not yet functional, as I still haven't
written a test driver for the library code yet.  In addition, the
code generation/rewriting code is not even in a testable state
yet, and is so early-stage it is nonfunctional.

The idea for this is to create a library that disassembles input
bytes and generates code that is a rewritten, safer version.
However, the actual policies for the safer code may differ
depending on use case, and the details of this will change as
requirements may change slightly.

The primary design goal is to generate code enforcing
SFI and CFI policies as quickly as possible, to allow for
on-the-fly secure instrumentation of dynamically generated code.