Added relocation structs to allow patching of addresses
After doing the initial rewriting, all statically identifiable
offsets in the code must be updated to reflect the new code layout.
This includes all direct control flow instructions, which need
to be patched to jump to the correct new destination.  However,
this cannot be done in the first pass because we don't know all
possible targets before we have generated the final code.  In the
previous iteration of this, I simply did two passes, the first of
which was only used to generate the mapping, and the second of
which patched the addresses.  However, for efficiency we do not
want to do all the duplicated work of analyzing all the
instructions twice, so instead I am storing the data needed to
make an efficient second pass.