Added code generation for jcc instructions, executed generated code
The sample code I am testing with is actually quite simple in terms
of control flow.  It's a better test of superset disassembly than
of rewriting instructions, because the actual control flow in the
code uses a single conditional jump instruction.  I need to write
a more complex test that uses a lot more edge cases for control
flow.