Changed driver to load from file, fixed chunk alignment crash
I wrote a more complex assembly example that is now loaded from
a file, but right now I have only tested it using the same function
I had tested in the file.  However, loading and rewriting from a
file is working, and it led me to find some problems with the way
I was adding nops to align chunks.  While it no longer crashes, I
will need to check that the alignment is correct.  Also, I changed
the signature of the code generator to accept architecture-size
independent arguments, so I hope that it will be easier to work
with both 32-bit and 64-bit code in the future.