Fixed call instruction alignment, changed generation to ret mapping
When testing the driver, I changed the base address and offset of
the code to be executed.  This caused an incorrect offset to be
executed at first, but it showed that call instructions were not
being generated correctly, and that their offset was not correctly
aligned on chunk boundaries.  I fixed this.

Also, I needed the mapping to determine what offset to call in the
driver, so I changed the code generation to return the mapping
instead of the code.  Since the code is not malloc'd anymore, there
wasn't really a reason to return it.