Added masking of return addresses
Now return addresses are masked off before the return instruction
executes.  This means that unless the calling code in the driver
happens to be aligned, the program will segfault when the rewritten
code attempts to return to the driver.  Since I'm really just
testing the correctness of the rewritten code, I haven't fixed
this.

Another cause for a crash is that the plt/got in the test binary
do not correspond with those in the driver, so a call to printf
will not work: the starting address of the plt isn't even the same
between the two, and the driver has more entries than the test
binary.  Obviously this data should match for actual generated
code that we would use this on, and this is just a way to generate
code that can be thrown at the rewriter.  However, it does mean
that a lot of parts of it will crash even when the rewriter is
working correctly.  For testing purposes, it is possible to jump
to the correct target in gdb, and this has worked.  In the future
I may want to create test code that works without crashing in the
driver.