Moved padding gen to function, now called once new inst size known
Prior to this, padding was generated based on the original
instruction's size, meaning if the new, rewritten instruction was
made longer or was composed of multiple new instructions, it would
be possible to generate instructions that spanned across a chunk
boundary.  Now, padding uses the new instruction length.  This does
mean that right now I don't plan to split new sequences of
instructions (which could reduce padding, but there are
restrictions, such as the fact that the masking of a control flow
target should not be separated from the control flow instruction).
This does have the side requirement that for now generated code
must be smaller than the chunk size.  If I need to make code longer
than that in the future, I must find a safe place I can split the
generated code.