Added alignment padding for predicted jump targets, unmatched insts
I have finally made the is_target callback functional.  It is now
used to ensure that potential indirect jump targets are aligned
so that masking off addresses for indirect jumps will work.  For
now, the actual is_target function that is passed to the rewriter
is very basic and matches just one function prologue, but this
is something that will be able to expand and adapt without having
to change the rewriter.

I also found a bug in that instructions that were not matched to
some control flow instruction did not receive padding, so they
could be misaligned.