Added trimming of invalid sequences from generated code
Any sequence of instructions that can directly flow into an invalid
instruction (i.e. an instruction not recognized by the
disassembler) can by extension be considered invalid, as the result
of running such code is unknown to the rewriter.

In addition, for benign code, we should be able to safely assume
that the program could potentially both take or not take a
conditional jump, so it is safe to assume that if code may fall
through to an invalid sequence, the sequence before and including
the conditional jump is invalid as well.  This also applies for
call instructions, as execution can reasonably be expected to
return from most calls, and for those that do not return (such as
the call to __libc_start_main in _start), there should be a valid
sequence afterward (In the case of _start, there is a hlt
instruction after the call, so it may prove necessary in the future
to record safe locations at hlt instructions).

This means that we can record the "last safe" location in the
generated code at each jmp or ret instruction, and if an invalid
instruction is later encountered, we can just roll back to this
last safe instruction, since it is guaranteed to jump somewhere
else.  This saves some bytes, as it makes the generated code
shorter.  Unfortunately, it appears that invalid sequences tend
to be fairly short, and not too many bytes are saved compared to
the total size of the generated code.  However, this still does
save a little space, and it doesn't leave clearly unintended,
useless sequences in the generated code.