Got basic indirect control flow functionality working
So far I have tested only on one type of instruction, (indirect
jmp to ecx), but it does correctly work.  The target destination
is patched in the original text, and the inserted code does appear
to check whether the destination is in fact in the old code
section.  The result is correctly retrieved, masked, and jumped
to.