Added additional test and driver, which does not work yet
I have now added a test that is a .so file, allowing for easier
testing of code, since code in a .so will know the correct address
of library functions.  However, there are still some issues that
I need to address before the test will work.  For one, I do not
support PIC, so any PIC will fail.  Also, there is something
causing a problem with the indirect control flow test case that I
need to figure out.

Also, right now the library sets the original code to not be
executable after rewriting it, but the destructor for the library
is automatically called on program exit.  This causes a segfault
at the end of program execution.  This isn't really a problem
right now, but it is a bit bothersome.