Added special case heuristic for PIC in test example
Now the PIC in get_fstring_c works with the rewriter.  There are
a few caveats:
	-This currently is dependent on the code using
	 __x86.get_pc_thunk.cx, which is used only when the ecx
	 register is free.  There are other variations for other
	 registers, and my implementation does not look for these.
	-There may possibly be unusual cases of bytes for a mov
	 from esp to a register before a ret appearing not in the
	 context of a thunk.  However, I can't think of a situation
	 where it would not be for the purpose of a thunk, so this
	 may or may not be a problem.
	-Any time a function from the test is referenced externally
	 it makes internal calls to the function instead of calling
	 the plt entry for that function. This is for purposes of
	 function pointer consistency, apparently.  Anyway, this
	 may not be a huge problem since it's a problem specific
	 to the fact that we don't have a modified plt/got to
	 smoothly redirect these functions.  We don't have this
	 because this is rewriting only a tiny bit instead of the
	 whole binary.  I do think that once the plt is
	 rewritten to handle indirect jumps and do a lookup,
	 this *should* happen transparently!
Also, any time I make a change to the test binary, I need to
change the special cases for get_msg1, get_msg2, and the entry
point.  Of course, the pattern for jmps to these labels is not
standard, and we may not expect a rewriter to identify it.  I'm not
sure, but target identification is still out of scope right now.