Added prototype code for mmap/mprotect and segfault handler
This code is adding the infrastructure to actually perform
rewriting on dynamically generated code.  However, there are many
parts that don't work yet and will have to be postponed for
future implementation, making this part of the implementation
a lot of proof-of-concept placeholder prototypes that won't
actually work for anything beyond the test driver.

The code implemented right now has partial support for loading the
library into arbitrary binaries (but the library's version of
libc does not seem to be completely position-independent, so it
still can't be loaded), hooking mmap/mprotect to remove the
EXEC permission (removing the permission with a shim works, but
I still don't actually hook the function, since this would
theoretically be done by the separate static rewriting component),
and providing a handler for SIGSEGV (partially implemented, but
not working yet.  This is what will need to call the rewriter).

Right now the current test driver I am using is driver3.  I intend
driver4 to be used when I can actually load the library without
relying on the dynamic linker, but right now that causes a
segfault in the library.