Implemented redirecting control flow after segfault handler
Now the handler can successfully redirect control flow to an
arbitrary target after it returns.  For now I simply am redirecting
to a ret instruction, so any non-executable pages that the program
attempts to execute will simply be returned to the caller.  Since
the code_caller is currently expected to return some addresses,
the later call to printf causes a crash when the code doesn't
return the expected values.  However, the important part is that
I can change the eip register in the calling context to anything
I want.

The mcontext_t struct in the handler is supposedly opaque, in that
it is machine dependent.  Obviously we are designing this for a
specific architecture, but it does mean that accessing fields in
it is outside what would officially be supported.  A result of this
seems to be that I can't figure out how to get the actual
definition of REG_EIP, which contains the index of that register
in the machine state.  I managed to change that register by looking
at the definition (found in /usr/include/sys/ucontext.h) and
finding the index manually, but I would prefer to avoid a magic
number if I can help it.  The definition in that header is
in an #ifdef that may be false, because I cannot access the
definition of REG_EIP even after including that header.