Moved rewriting functionality into segfault handler
Now rewriting occurs when there is a segfault; the original page
that a target was in is rewritten, and then all future calls to
that original page are automatically redirected to that page.

This does currently only support rewriting a single page once,
with no way to rewrite the same page or more pages.  I need a
data structure to track pages that have and have not been
rewritten.  I will want to track non-rewritten pages for safety
because only pages that the original program *tried* to set as
executable should be rewritten.

I made changes to driver4 at first because I got it and driver3
mixed up; it doesn't matter because driver4 would have needed these
changes anyway, but I just wanted to note that driver3 contains the
prototype that is testing the rewriting functionality right now.