Added new driver test to also test mmap/mprotect hooking
driver5 maps pages, writes code in them, and tries to call the
code.  By hooking mmap and mprotect, the driver is prevented from
setting the pages executable, and the segfault handler interposes
when it attempts to execute the code.  It rewrites it and jumps to
the rewritten code.

There is an additional edge case I found that makes the rewriter
not work for rewriting a single full page.  If it tries to read
an address right at the end of the page, it may try to read bytes
on the next page, which may not be mapped.  Right now I allocate
an extra page, but this will eventually need to be addressed in
one way or another.