Changed segfault handler to allocate new code at random address
Now the rewritten code is not placed at a fixed address.  Instead,
mmap is passed a void pointer initially, which allows the kernel
to choose any address it wishes to allocate the pages.  The only
downside is that we don't know how many pages we should allocate
up front, and providing too small a size may result in the kernel
placing the allocated memory right next to another allocated
region.  If we predict too low a size and end up writing past
the original size, we may overwrite bytes in another region,
causing a segfault.

For now, I allocate 4 pages for the single page of code I rewrite,
but not only does the rewriter not know this (and do unnecessary
additional mmaps), this is a hackish way to do it.  Eventually we
will likely be writing more than a single page at a time, and we
will need to scale things based on the number of input pages.
Also, we may want to de-allocate pages at the end after rewriting
is done if our estimate is too high.