Fixed issues with pic generation, gets further before crash
Now some of the problems I was dealing with are solved.  PIC now
does seem to work, and chunks of the library can run without issue.

I also encountered a crazy issue with musl in which the definition
of a specific data type (off_t) differed between musl and the sys-
tem headers, resulting in calls to mmap from the main application
passing a 32-bit integer to musl's implementation, which expected
a 64-bit integer.  This difference was not detected by the linker,
and the function would read garbage bytes on the stack as it
attempted to read 32 extra bytes on the stack.  I have solved this
by hacking musl to define off_t as 32 bits, but I would like a
less hacky solution.

Unfortunately, deep within Capstone, trouble arises.  Its antics
with dynamic memory allocation and perhaps other unknown
complexities are still causing crashes, and I have decided to try
using a completely different disassembler.  I have now modified
the superset disassembler to work with a simpler disassembler that
doesn't use dynamic memory allocation, and the next step is now to
convert miniverse as well.  Hopefully this will finally allow
easy static linking without using the linker.