Started implementing new approach to loading library
After some investigation, I have found that even with PIC flags,
a binary is still laden with relocations that must be resolved at
load time.  This makes simply dropping an arbitrary binary blob
into an address space infeasible.  udis86 keeps tables of data for
disassembly, and at least one table contains pointers to other
tables.  These pointers must be updated to reflect wherever the
binary ends up being loaded, and I think there will probably be
relocation entries for any case in which data refers to other data.
Therefore, even if we can eliminate absolute addresses in code, we
may not be able to eliminate absolute addresses in data.

However, I have found that the generated binary now only has a
single type of relocation entry, R_386_RELATIVE, and I have decided
to simply parse the relocation entries myself when I load the lib
into memory.  Implementing this will allow me to either dynamically
load the lib at runtime, or statically patch the lib and place it
at a fixed address in a binary.  This still needs to be implemented
and is the next step I intend to do.

I also found that the data segment in the library was being loaded
at a one-page offset from the text section when compared to its
offset in the file.  This offset meant that any data references
were wrong if I just tried to map the whole file in at once.  I
have changed the loading process to parse the elf header and map
in LOAD segments to their requested addresses.  I also zero out
the bytes in the .bss section (manually, because .bss starts right
after .data not at a page aligned address so I can't use mmap)
because musl will try to read an address from .bss that would
normally have been initialized at some point and uses that to try
to perform system calls if its value is not zero.  When I map the
data segment, the region that would contain .bss values actually
contains bytes from sections like .comment and debug sections
since .bss doesn't take up any space in the file, so I clear the
bytes.  If anything relies on a value here or in .data that is
set during initialization, we may have other issues in the future.