Added start of binary patcher to insert miniverse lib in binaries
I still need to decide on a "least destructive" way to modify phdrs
in an arbitrary binary.  My initial plan was to temporarily
overwrite the start of the text section (and restore it at runtime)
but I had an idea that I know I partially attempted previously by
attempting to shift content forward but load it at a lower address
corresponding to the shift.  My previous notes indicate that this
required changes to the loader, but I can't recall why.  If the
reason is impossible to get around (and it might be, I remember
a lot of frustration with the loader), then overwriting the text
section will work as long as its large enough.  I got it working
before, but what's unappealing is that it requires making the start
of the text section writable for a brief period on start.  Perhaps
not too bad, I'm not sure.  We already have code that sets regions
from writable to executable, so I don't know if this really adds
more "exploitable" code or not.