Added first draft of entry point assembly
The assembly hasn't been tested and almost certainly won't work
correctly the first time.  It is supposed to map in a memory
region for the miniverse library and then copy its bytes from
within an already-loaded segment in the binary to the intended
load address of the library, then jump to the library, which can
restore the original entry point and jump there after it performs
initialization to resume normal execution.

Dr. Hamlen had the suggestion to not move the phdrs at all, and
instead patch over the original entry point directly and expand
the last LOAD segment (at the highest address) to hold the data we
need.  Since anything that's not a LOAD segment should not be
available to the process without opening itself for reading
(well, hopefully; it's correct as far as I can tell), it should be
safe to clobber any data following that last segment.  This also
means that we don't need to worry about opening the file for
reading again, as we will have the data we need already mapped in,
even if it's in a segment that isn't executable.  As long as the
last segment is readable, we can just copy the bytes we need from
there to new memory regions we map at runtime, using the assembly
we paste over the entry point.  This should be very universal for
elf files and doesn't alter the program headers at all.

We could also try to unmap the bytes we added on the end of the
last segment after we're done if we want, although I think that's
probably unnecessary.