Continued implementation of binary patcher
The code has gotten complex, and Dr. Hamlen suggests trying to
instead implement a solution that can be tested without having
to go through the binary patcher to debug changes, so that a
simple test could link the library in, instead of having to run
everything through the patcher first.

I have been considering the options for a simplified approach,
which would make things less universal, but make it easier to
implement.  The interface to the dynamic linker appears to be more
complicated than I feel it should, but even using the dynamic
linker feels like overkill when I just want to load my library into
memory and have it just work without dealing with the linker.

One possibility would be to attempt to start a parent process with
our library already loaded, and then somehow keeping the library
resident even when we exec the child process.  It does appear that
it may be possible to use clone to maintain some shared memory
regions, so that could be an avenue to keep the library loaded in
the child.  However, the more I investigate, the more it appears
that the exec syscall completely wipes out everything when loading
the new binary.

Another option is to take the loaded library in the parent and then
try to modify the address space of the child to map the library in
after the child has exec'd.  Unfortunately, that means that somehow
the child must wait until the library has been mapped into its
address space before it can initialize it, and it's not clear what
the best method for determining whether a region of memory is
mapped or not is.

A third option that I am considering as the best possibility right
now is to try changing the library to be position-dependent, and
then mmapping it to a fixed address.  If it's statically-linked and
at a fixed address, perhaps we won't need the linker.  If not, then
I do also have a fallback solution of dumping the memory image of
my patched version of the library to a file, and mmapping that.

One issue is that I do need a way to initialize the library, so I
will need to patch over the entry point with inline assembly either
way, and this means I still need to have code to restore the entry
point of the binary, meaning I still have some difficult-to-debug
initialization code regardless....