Got patched binary to successfully mmap flat binary image of lib
I have not tested whether the mmapped lib works yet, and I also
have not written the code for restoring the original entry point.

Right now I assume the entry point code I overwrite the original
entry point with to be exactly one page, but this is excessive and
could be a problem for tiny binaries.  I should eventually only
copy the minimum amount of bytes needed.

Dr. Hamlen suggested that I get a statically linked implementation
of dlopen working so that I can then load basically anything after
loading that.  Doing that still requires this first phase of using
mmap to load in that module first (and also still requires that I
write code for restoring the original entry point), so I can work
on patching the entry point first.  Right now I am still trying to
load my already-patched library instead of a dlopen library because
I am worried that getting a statically-linked dlopen to work will
take some more effort than it might first appear.  Musl has an
implementation of dlopen, and a brief inspection shows that it may
intentionally disable dlopen functionality when statically linking
musl.  This does intuitively make sense, since the dynamic linker
does not need to be present for statically-linked binaries, but
it means it's hard to just load a self-sufficient binary blob into
memory and call linker functionality.  It needs to know where the
linker is in the host binary, and if it's just mapped in, it does
not undergo any initialization (where the linker would run first
and set a global somewhere it knows about).  However, this is the
part of dynamic linking I still need to study up on: how does the
dynamic linker communicate with libc, i.e., how does it tell libc
where it's loaded and what calls you can make into it (if that is
even how it works)?