Got patched binary to call mapped library code
Now the patched entry point maps a library from a file and jumps
to its entry point, which is specified in data placed alongside the
patched code at the binary's entry point.  The library could be
anything from the point of view of the entry point assembly.

The library still needs to be able to restore the original entry
point, but I have noticed it isn't even capable of printing out
debug statements right now: musl's printf function eventually
writes to a pointer to __stdout_FILE in the library's data section,
which I did not map as writable.  I either can map the library as
writable+executable, which is dangerous but could work in the short
term, or pass an argument to it giving the memory range of the data
section and perform an mprotect call right after entering the
library.  As long as mprotect doesn't write to the data section,
it should be safe to execute that before .data is writable.  It's
not clear whether it would be needed to have .data writable for
a simple dlopen library or not, so I'm not sure whether I should
try adding this yet.