Changed assembly to pass correct args to library code
Changing the loaded library to W+X works for now, and it now does
successfully print out the arguments passed to it.  We obviously
want to avoid leaving the library W+X, but perhaps it could fix
that by itself.  The next step is to save a backup of the entry
point, restore the entry point, and return to the entry point,
allowing the original program to run as it did before, except with
the difference that it has a library mapped in its address space.