Got library code to restore entry point and return to it
Now the library successfully restores the entry point assembly and
returns straight to the original entry point.  As a future step,
I can set up the segfault handler before returning to the original
binary and then my library can handle all segfaults generated by
the binary.

One thing that still needs to be fixed is that the stack is not
cleaned up properly when returning from the library.  I think this
clobbers argc and argv, because those are expected to be at the
top of the stack.  I need to force my library to do cleanup for the
caller, violating normal calling conventions since the target of
the return (the entry point) doesn't know we are returning to it.
This is only a problem because I wanted the library entry point to
be in C instead of assembly.  This could be solved by having an
assembly entry/exit wrapper, but I think I can solve this by just
implementing my own function epilogue in inline assembly at the end
of the function.