Mprotect only one page in miniverse_entry, changed debug text
Binaries produced by upx are quite small, so attempting to change
permissions for even two pages failed, as there was no page mapped
after the first page.  I think I can't just settle with calling
mprotect for one page in all cases though: the entry point in many
binaries will not necessarily be nicely page aligned, and if it
isn't, the entry point code I patched over it may span onto another
page.  Also, for now I restore an entire page's worth of the
original entry point, so if it's not exactly page aligned it should
segfault when I try to restore the entry point.  I'm not sure of
the most reliable way to determine how much we need to mprotect;
we may be safe with 2 pages if the entry point is not page aligned,
but there will still be potential binaries that violate that
assumption.  At minimum, the original binary must have enough space
to fit in our new entry point code, so we may be able to calculate
whether that length crosses a page boundary.

This has revealed that the loaded library does not work for ALL
cases, as it will segfault when attempting to read the location of
errno if an error occurs and it wishes to record it (but it is
unknown whether this will happen in all binaries or just some, such
as statically linked binaries---I just know that it happened in a
statically-linked binary compressed by UPX).  When I fixed the call
to mprotect, this issue went away because the call succeeded and so
it didn't set errno.  However, this could be a harbinger for future
problems if a library call ever fails for any reason.

I have also started trying to get the library working for luajit,
compiling the library in from source, but there are some issues
I need to address with supporting multiple rewritten chunks of
code (luajit appears to have numerous mmap/mprotect calls when it
generates jit code), and I can't expand the memory region for new
mapped code as I currently am (mmapping a new page at the end of
my currently mapped region), as the kernel can give us a page
directly before an already mapped region for our initial page; I
need to switch that code to use my pagealloc library instead.
There might be some overhead from having to copy all the memory
over to a larger region, but the main concern is to make sure I
don't store any pointers into the original region, since a realloc
will necessarily change where that memory is mapped.