Changed library to use page_realloc to alloc new code pages
Now in theory it should not be an issue expanding the memory for
the new rewritten code, and tests with driver5 do seem to show
it to work.  However, after making all my changes it does not seem
to need to expand the size of the rewritten code at all.

I still need to make changes to the segfault handler so that it
can handle multiple rewritten memory regions, which should take a
bit more work than this.