Renamed mapper.c to miniverse.c
I'm doing a small amount of clean-up and fixes to the rewriter,
since it works but has many unfinished TODOs.  I have to decide
what to try to get working next: another jit-compiler or upx?

I have now confirmed that rewriting does in fact work for generated
code in luajit, which is very promising!  However, there does
appear to be a serious performance hit, and I have been trying to
investigate the primary cause of the slowdown.

I briefly experimented with valgrind to try to profile luajit, but
valgrind breaks my rewriting, as my segfault handler is never
triggered.  Somehow my changes to the permissions for mmap and
mprotect don't work, as the program appears to run fine using the
original generated instructions I set as non-executable.

One possible cause of high overhead, besides the obvious overhead
of my debug print statements, is from the segfault signal and the
cost of switching context to and from the segfault handler.  I'm
not sure how to avoid the overhead of that without rewriting the
code calling the generated code region to instead call my
rewritten code instead.  The complexity with that is that every
time the original code region is updated, I would either need to
change where the calling code jumps to again or have it jump to a
shim that checks whether we need to rewrite the code or not.
Either way, we won't know ahead of time where all the call points
are, so this would have to be just a run-time optimization.