Disabled debug output with ifdef guards
I'm still doing some cleanup as I prepare to try to get another
jit compiler working.  I have been working on compiling the
javascript engine used in Firefox, and now it is ready for me to
try to add miniverse to it.  However, the fact that I have to
pass a modified handlers.c to luajit means I want to refactor a
bit first.