Refactored register_handler to accept is_target callback
By passing a callback for is_target to register_handler, code
using miniverse can specify its own custom callback or pass NULL
in order to use the built-in default.  Before, there was no way to
use miniverse without defining your own is_target function, which
required that code to be added somewhere to luajit.  I ended up
putting it into a copy of handlers.c in the luajit source.  Now,
I have placed an extern declaration or register_handler in
miniverse.h so that the only thing external code needs to do is
include miniverse.h and call register_handler (and be compiled
with linker wrappers for mmap and mprotect and the required
libraries).  This makes it so that adding miniverse requires
almost no source code changes.  Of course, we do want this to
eventually be added to binaries, but this makes testing much easier
by making changes to miniverse and libminiverse's inclusion in a
program as straightforward as possible.