Wrapped most printfs in DEBUG, support ret imm16, changed mask val
Now when DEBUG is not set, less debug statements are produced.
There are still a few, but I plan to either remove them or also
wrap them with #ifdef DEBUG.  It is true that there is honestly
too much debug output, so I probably should remove some entirely.

I have also added support for ret instructions that pop an
additional imm16 bytes off the stack.  This was needed for some
javascript jit code.

I also changed the mask for indirect jumps, because the top of
the mask masking off the top few bits was causing problems due to
our generated code being placed at a very high address.  This needs
to be investigated further to determine if there is a need or use
to mask of top bits in any situation.  My original recollection was
that this would prevent the generated code from jumping into the
VDSO or linker, but I'm getting addresses from mmap starting with
with the top 4 high bits set, so it doesn't seem possible to just
mask off these bits.  Besides, some legitimate code may call into
those high addresses, so I can't recall the reasoning for the mask
starting with 0x3F vs 0xFF.