Handle modifications to subregions, ind jmps between jit regions
Now we split a region into two if a change in permissions occurs at
the end of the region.  I don't actually support splitting a region
into three parts yet, but I have not encountered this yet.

I also encountered a problem with indirect jumps occurring between
jit regions in which the destination had not yet been rewritten.
This resulted in a lookup failure, as the target destination was
not in a rewritten region.  This was solved by rewriting regions
as soon as the binary attempts to set them to R-X from RW-.  This
is a heuristic that won't work for RWX regions, but it appears that
the jit code I have worked with so far is well enough behaved to
not need to worry about that yet.

Other minor improvements were made as well, including some
refactoring, cleanup, and small fixes.