Added optional lookup code for return instructions
It turns out that some return addresses in some cases may be
placed by the code generation instead of a rewritten call
instruction, and this means that rewritten returns that simply mask
off the return address will make the program attempt to jump to
a bad offset in the old code section.  Therefore, I have added code
to have return instructions perform this lookup.  This must also
be paired with identifying all instructions after a call
instruction as a jump target, so that their lookup entries will be
properly recorded in the old text section when rewriting.  I did
this by saving a variable marking whether the previous instruction
was a call.

The new generated code works on my simple driver tests now, but I
have not tested it on anything more complex yet, which is the next
step.