Back up and restore original code if region reverts to writable
It turns out that luajit does not rewrite everything in the jit
code region it allocates when it makes changes, which makes sense.
However, this caused problems because I am patching the original
code with mapping entries, and if the program makes changes to the
original region without refreshing it completely, I will rewrite
the patched version of the original code, which results in invalid
code and broken program behavior.  The solution I have implemented
for this is to save a backup of the original code right before
rewriting it, and then restoring the backup if I detect that a
region's permissions are being changed from X to W.  This will
not work for regions that are permanently RWX, but there are
probably quite a lot of reasons that RWX regions will not work with
my implementation by now.  This does appear to work well for luajit
now, and performance is still better for many benchmarks with the
jit turned on, so results are still good for luajit.

I also made sure that reaching the end of an instruction sequence
will reset whether the previous instruction was a call.