Restore bytes for backed up subregions, and allow region expansion
Now when we detect that mprotect is called for an address within
a region, we just restore the bytes for that subregion.  The region
will be split when the program tries to make it executable again,
and the resulting new region will then be backed up.  I now also
allow for a region's size to be expanded instead of it making a
new larger region with the same start address, which was a bug.

I should make sure that I free region backups to avoid memory
leaks, since every time a region is rewritten again it should have
a fresh backup with the old backup freed.

I also created a new test for the changes I made to ret
instructions, as well as for modifications to subregions.