Fixed some memory leaks
Now both stale code backups and stale rewritten code are freed.
There are some complexities in jit code that mean that it may not
be possible to actually free the rewritten code, as there could be
a return address to the old rewritten code even after generating
the new code, but that still will require some sort of special
handling if I encounter it.  The primary issue is that I'm only
freeing old rewritten code for a region if it is being rewritten
again, so that means that pointers into the stale region need to
be redirected into the fresh region in order to be safe.  For now,
instead of adding support for that, I'm just getting rid of the
region entirely, and hopefully this will make detecting code that
may have issues with this easier in the future (as long as a new
region isn't allocated in its place, it will segfault immediately).