Added new test and patcher script to make compatible targets in bin
I encountered a problem that I'm surprised I didn't encounter
earlier: indirect jumps or returns to the original binary usually
fail!  Control flow into rewritten code work, but since the
rewritten code performs a lookup and a mask, an indirect control
flow back to the original binary would only work by chance if the
target was both 16-byte aligned and its first byte happened to not
have its lowest two bits set.

Now I have implemented a small script to help with tests in which
we assume we can build from source.  Inserting a call to the script
as part of the build process and changing it so that I compile each
source file to a .s file allows me to patch the assembly for each
file before its offsets are fixed.  Along with this, I implemented
a new test, driver9.

I pad every function entry to make them 16-byte aligned, and insert
a nop at the start of every function.  This allows rewritten code
to indirectly jump to or call any function the original code.

I also align a subset of indirect call instructions (matching
certain patterns I have noticed are commonly used) such that the
instruction after the call is 16-byte aligned.  This allows
rewritten code to return to the original code. This also removes
the need I had before to use the code_caller hack to force the
instruction after the call to be aligned.  However, in my most
recent tests I had forgotten to insert a nop directly after the
call instructions.  I have just noticed this, because it did not
cause any issues in any of my testing.  I think I may have just
discovered a bug in the code I generate for ret instructions, so
I need to investigate this next.